HPCC-13042 Set WsTopology.default_client_version to 1.21
Existing ECLWatch UI uses the default_client_version to
access ESP service. The WsTopology.default_client_version
has to be set to 1.21 in order for ECLWatch to get the
replicateOutputs option.

For now, if a SOAP client does not specify its version
inside a SOAP call, ESP will use 0 as the version. Therefore,
incrementing the default_client_version will not break SOAP
client even if the SOAP client does not send its version.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>