HPCC-13113 Fix incorrect use of default_client_version
Inside ESP ecm files, the 'version' variable defines the
latest version of an ESP service. The 'default_client_
version' variable may be used to define the version of an
HTTP ESP client when an ESP HTTP client does not specify
its version. In this fix, the 'default_client_version'
variables are removed from the existing ECLWatch ESP
services since, until now, ECLWatch HTTP forms always use
the latest version of an ESP service and the 'version'
variables have already been defined in the ECLWatch ESP
services. (The 'default_client_version' variable may added
whenever needed later.) The hidl code is modified to use
the 'version' variable as the version of an HTTP client
if the client does not specify its own version and the
'default_client_version' variable is not defined.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>