HPCC-13546 Add filtering to serializeLockData()
In HPCC-13545, the serializeLockData() is created for
serializing lock data for both getLockInfo() and new
getLocks(CMessageBuffer &mb). In this fix, the optional
filtering is added into the serializeLockData(). Using
HPCC-13545 and HPCC-13546, filtered lock data is
retrieved from dali without formating. The formating
is done in the client side.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>