HPCC-9999 Release log queue lock after a queue item is read
In the existing code, after a log thread gets log queue lock, it
does not release the lock before finishing all of queued log items.
ESP logging manager may wait for a long time to add a log request
into the queue if the log thread is working on the queued items.
In this fix, the log thread releases the queue lock as soon as it
reads a queue item from the queue.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>