HPCC-19386 Reset Transaction Seed using pre-configured conditions
A user of ESP transaction logging asks for resetting transaction seed
after transaction ID reaches a limit of the ID length. ESP transaction
logging should allow a user to pre-configure a logging agent to reset
transaction seed and sequence number in order to meet the limit of the
ID length.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>