HPCC-18655 Multiple changes
The changes are added based on Gavin's comments. ESP only caches
ESP responses if there is cache configuration. An ESP may be
configured to eithor have one default cache server or multiple
named cache servers. ESP may not be started if the cache server
cannot be accessed. To avoid runtime delay, ESP cache clients
are created when ESP starts. To avoid overhead on systems that
do not have memcached installed, a quick memcached available
check is added. The cache clients are moved from EspHttpBinding
to CEspServer because multiple EspHttpBindings may share one
cache client. The cache client may be accessed from outside of
EspHttpBinding.

The ensureESPCache configuration is removed since ESP may not
be started if the cache server is configured but cannot be
accessed. Also cleaned some code.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>