HPCC-12757 Support HTTP Content-Encoding header
In HTTP call, a client may send gzip compressed requests
with the Content-Encoding header. ESP should be able to
read the request and, if possible, send back the response
with the Content-Encoding header. ESP should also support
the HTTP Accept-Encoding header.

The zlib_deflate() is changed to reallocate the output buffer.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>