HPCC-22208 Ensure correct file path for WsFileIO
In the existing WsFileIO code, absolute file path is set using
dropzone folder + relative path defined by user. When a dropzone
folder is configured as '/' and the relative path is started
with '/', createIFile() fails because of the double '/' in the
absolute file path. In this fix, extra check is added to ensure
correct file path.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>