HPCC-22136 Revise based on review from old PR
1. Remove Workunit API importWorkunitFromZAPFile(); Move the
wuFactory->importWorkUnit() to WsWorkunits service so that we
do not have to call the getWorkUnitFactory() and get ESP DATA
directory every time  when importing a WU;
2. Add getSlaveLogFileName() which returns the name of
slave log file for ECL WUs, imported WUs, and different legacy
WUs. Three WsWorkunits functions are revised to use the
getSlaveLogFileName();
3. depr_ver the log date and ClusterGroup from ThorLogInfo.
The existing getWUThorLogInfo() still contains the conditional
code for retrieving log date. New getWUThorLogInfoLW() is
added to return the ThorLogInfo without log date and group name.
From this version on, getWUThorLogInfoLW() should be used
by WsWorkunits to get the ThorLogInfo;
4. Change addProcess() to handle @max and @pattern;
5. Add getWUAttributes() to collect the DistributedAccessToken
and Scope from the newly created workunit. The information is
added into imported WU xml. In the setImportDebugAttribute(),
remove the user string and format the DebugValue to be more
easily parsable;
6. Add code to remove the workunit xml, progress xml, and the
uploaded ZAP report after import;
7. Change the readWUXMLFileToPTree() so that it is also used for
the GraphProgress;
8. Change getSlaveLogPattern()/setSlaveLogPattern() to
getProcessLogPattern()/setProcessLogPattern().
9. Replace the logSingleFile flag with dedicatedLogFiles. Change
logSingleFile() to usingDedicatedLogFiles();
10. In WsWorkunits code, merge 2 getWorkunitThorSlaveLog() into
one;
11. Other minor changes.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>