HPCC-23880 Establish contention on /Sessions/Process if failed
1. When esp is starting or attached to the dali, retry the SDS
connection to /Sessions/Process 6 times if failed;
2. When cleaning ESP session data, retry the SDS connection to
/Sessions/Process if failed;
3. Rename the ensureSDSSessionDomains() to ensureSDSSession();
4. Inside the ensureSDSSession(), the code for reading Session
Domains settings is moved into new readSessionDomainsSetting().
It is called only esp is starting;
5. The ensureSDSSession() is called in 2 scenarios: 1. after
the readSessionDomainsSetting() is called and if esp is not detached
from dali; 2. when esp is attached to dali if the ensureSDSSession()
has not being called;
6. In the existing code, the EspHttpBinding::setSDSSession() is
used to ensure that /Sessions/Process has an Application branch
for the esp binding. It is called when EspHttpBinding is created.
That does not work if esp is not attached to dali when the
EspHttpBinding is created. The related code has been moved into
the ensureSDSSession() so that the code can be called when esp is
attached to dali;
7. Also rename querySDSConnection() to getSDSConnection() and rename
querySDSConnectionForESPSession() to getSDSConnectionForESPSession().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>