HPCC-25484 Move daliadmin helper functions for ESP WsDali
Create 2 new files in dali/daliadmin: daadmin.cpp and daadmin.hpp.
Move the code from daliadmin.cpp to daadmin.cpp except for main(),
usage(), and a few lines needed by main().
For the helper functions used inside the main(), add the 'extern
da_decl' declarations into the daadmin.hpp.
For those helper functions in daadmin.cpp, remove the 'static' and
the default input values if any.

This is the 1st commit. This commit does not have new daliadmin.cpp
which will be in the 2nd commit.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>